240319
Aim of the src files to satisfy the following needs:
Design a classes Tender, PWD and contractor. Use an inheritance and association
relationship appropriately. Write a java program to store the tender details by the
contractor.
2. Design an interface tender. PWD and contractor implements the tender interface. Write a
java program to store the tender details by PWD and contractor from their perspective.
3. Design a Class contractor and store the contractor information using array of objects. Write
a java program to store the contractor details. The java program should have a mechanism
to avoid the duplicate entries of contractor object. If duplicate entries identified user
defined exception should be thrown to handle this error.